refactor(kerek): move kerek into its own file
