+++ EDITTOOL.CPP
@@ -10,6 +10,7 @@
 #include "M_PIC.H"
 #include "main.h"
 #include "menu_pic.h"
+#include "object.h"
 #include "pic8.h"
 #include "TOPOL.H"
