+++ EDITUJ.CPP
@@ -15,6 +15,7 @@
 #include "M_PIC.H"
 #include "main.h"
 #include "menu_pic.h"
+#include "object.h"
 #include "pic8.h"
 #include "platform_impl.h"
 #include "platform_utils.h"
