+++ LEJATSZO.CPP
@@ -8,6 +8,7 @@
 #include "LEPTET.H"
 #include "LGRFILE.H"
 #include "main.h"
+#include "object.h"
 #include "physics_init.h"
 #include "platform_impl.h"
 #include "platform_utils.h"
