+++ TOPOL.H
@@ -4,6 +4,7 @@
 #include "state.h"
 #include "vect2.h"
 
+class kerek;
 class polygon;
 struct motorst;
 
@@ -11,29 +12,6 @@ void set_gyuru_attributes(polygon* pgy);
 
 #define MAXKEREK (52)
 constexpr int MAX_VERTICES = 5000;
-
-enum { T_CEL = 1, T_KAJA, T_HALALOS, T_KEZDO };
-enum { KT_NORMAL, KT_UP, KT_DOWN, KT_LEFT, KT_RIGHT };
-
-class kerek {
-  public:
-    vect2 r;
-    int rxint, ryint;     // Kaja lejatszas kozbeni kirajzolasahoz
-    int rxint_v, ryint_v; // Ugyanez view-ra
-    int tipus;
-    int kajatipus; // Ha kaja, akkor lehet gravitacios is
-    int foodsorszam;
-    int aktiv;          // Lejatszas soran ha kaja, megvan-e meg
-    unsigned char szin; // Csak lejatszasnal kell
-    int sinfazisint;    // 0-tol 999-ig mehet
-    double sinfazis;    // 0-tol 2Pi-ig megy
-
-    kerek(double x, double y, int tipus);
-    kerek(FILE* h, int verzio);
-    void kirajzol(void);
-    void save(FILE* h);
-    double belyegszamitas(void);
-};
 
 #define MAXSPRITE (5000)
